refactor for multiple implementations
